refactor: package path of DTOs
